Reworked UsersWithPermissionInfoController
